# Optimize NewSafeID using atomic