--- a/messageID_test.go
+++ b/messageID_test.go
@@ -1,0 +1,27 @@
+package slack
+
+import (
+	"testing"
+)
+
+var id int
+
+func BenchmarkNewSafeID(b *testing.B) {
+	b.ReportAllocs()
+
+	idgen := NewSafeID(1)
+	for i := 0; i < b.N; i++ {
+		id = idgen.Next()
+	}
+}
+
+func BenchmarkNewSafeIDParallel(b *testing.B) {
+	b.ReportAllocs()
+
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			idgen := NewSafeID(1)
+			id = idgen.Next()
+		}
+	})
+}