Optimize NewSafeID using atomic instead of a mutex lock.

- [x] is it performance related

---

```
name                  old time/op    new time/op    delta
NewSafeID-20            14.0ns ± 1%     5.9ns ± 1%  -57.76%  (p=0.008 n=5+5)
NewSafeIDParallel-20    23.8ns ± 1%    21.0ns ± 2%  -12.08%  (p=0.008 n=5+5)

name                  old alloc/op   new alloc/op   delta
NewSafeID-20             0.00B          0.00B          ~     (all equal)
NewSafeIDParallel-20     8.00B ± 0%     8.00B ± 0%     ~     (all equal)

name                  old allocs/op  new allocs/op  delta
NewSafeID-20              0.00           0.00          ~     (all equal)
NewSafeIDParallel-20      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```